Added ability to initialise other arrays for direct class initialisation
This includes the charge, spin arrays as well as the various other flags like
have_ncm or the number of spins.

However, the only things that are required to directly initialise a class are
real_lat and the data array.

The other attributes are given 'sane'(-ish) default values, notably:
 * nspins defaults to 1
 * have_ncm defaults to False
 * is_den is False
 * Uninitialised attributes that contain numpy arrays such as spin and charge default to None.